api/pyproject.toml
[project]

name = "api"
version = "0.1.0"
description = "Fluent flow api"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "aioboto3==15.0.0",
    "aiobotocore==2.23.0",
    "aiofiles==24.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aioitertools==0.12.0",
    "aiosignal==1.4.0",
    "aiosmtplib==3.0.2",
    "alembic==1.16.4",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "argon2-cffi>=23.1.0",
    "argon2-cffi-bindings>=25.1.0",
    "asyncpg==0.30.0",
    "attrs==25.3.0",
    "audioop-lts==0.2.2",
    "azure-cognitiveservices-speech==1.45.0",
    "azure-core==1.35.0",
    "babel==2.17.0",
    "bcrypt==4.3.0",
    "black==25.1.0",
    "blinker==1.9.0",
    "boto3==1.38.27",
    "botocore==1.38.27",
    "certifi==2025.8.3",
    "cffi>=1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.6",
    "csvw==3.5.1",
    "dataclasses-json==0.6.7",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "dlinfo==2.0.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "fastapi-mail==1.5.0",
    "fastapi-users==14.0.1",
    "fastapi-users-db-sqlalchemy==7.0.0",
    "frozenlist==1.7.0",
    "grandalf==0.8",
    "greenlet==3.2.3",
    "groq==0.31.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-oauth==0.16.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "isodate==0.7.2",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.72",
    "langchain-groq==0.3.7",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.6.3",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.3",
    "langgraph-sdk==0.2.0",
    "langsmith==0.4.13",
    "language-tags==1.2.0",
    "limits==5.5.0",
    "makefun==1.16.0",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "numpy==2.3.2",
    "orjson==3.11.1",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "paddle-python-sdk==1.10.0",
    "pathspec==0.12.1",
    "phonemizer==3.3.0",
    "platformdirs==4.3.8",
    "propcache==0.3.2",
    "pwdlib==0.2.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydub==0.25.1",
    "pygelf==0.4.3",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "rdflib==7.1.4",
    "redis>=6.4.0",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rfc3986==1.5.0",
    "rich==14.1.0",
    "rich-toolkit==0.14.9",
    "rignore==0.6.4",
    "rpds-py==0.27.0",
    "ruff==0.12.8",
    "s3transfer==0.13.1",
    "segments==2.3.0",
    "sentry-sdk==2.34.1",
    "seqlog==0.4.3",
    "shellingham==1.5.4",
    "six==1.17.0",
    "slowapi==0.1.9",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "tenacity==9.1.2",
    "typer==0.16.0",
    "typing-extensions==4.14.1",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "wrapt==1.17.3",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "zstandard==0.23.0",
]
